Rename or move some simple tree source files and tests
