+++ packages/dds/tree/src/simple-tree/toStoredSchema.ts
