+++ packages/dds/tree/src/test/feature-libraries/chunked-forest/codec/schemaBasedEncoding.spec.ts
@@ -62,7 +62,7 @@ import {
 	getStoredSchema,
 	toStoredSchema,
 	// eslint-disable-next-line import/no-internal-modules
-} from "../../../../simple-tree/toFlexSchema.js";
+} from "../../../../simple-tree/toStoredSchema.js";
 import { numberSchema, stringSchema } from "../../../../simple-tree/index.js";
 
 const anyNodeShape = new NodeShape(undefined, undefined, [], anyFieldEncoder);
