+++ packages/dds/tree/src/test/shared-tree/schematizeTree.spec.ts
@@ -47,7 +47,7 @@ import {
 	type TreeViewConfiguration,
 } from "../../simple-tree/index.js";
 // eslint-disable-next-line import/no-internal-modules
-import { toStoredSchema } from "../../simple-tree/toFlexSchema.js";
+import { toStoredSchema } from "../../simple-tree/toStoredSchema.js";
 import { jsonSequenceRootSchema } from "../sequenceRootUtils.js";
 
 const builder = new SchemaFactory("test");
