+++ packages/dds/tree/src/test/simple-tree/api/typesUnsafe.spec.ts
@@ -4,10 +4,10 @@
  */
 
 // eslint-disable-next-line import/no-internal-modules
-import type { ReadonlyMapInlined } from "../../simple-tree/api/typesUnsafe.js";
+import type { ReadonlyMapInlined } from "../../../simple-tree/api/typesUnsafe.js";
 // eslint-disable-next-line import/no-internal-modules
-import type { numberSchema } from "../../simple-tree/leafNodeSchema.js";
-import type { areSafelyAssignable, requireTrue } from "../../util/index.js";
+import type { numberSchema } from "../../../simple-tree/leafNodeSchema.js";
+import type { areSafelyAssignable, requireTrue } from "../../../util/index.js";
 
 type MapInlined = ReadonlyMapInlined<string, typeof numberSchema>;
