+++ packages/dds/tree/src/test/simple-tree/toStoredSchema.spec.ts
@@ -7,7 +7,7 @@ import { strict as assert } from "node:assert";
 
 import { SchemaFactory } from "../../simple-tree/index.js";
 // eslint-disable-next-line import/no-internal-modules
-import { toStoredSchema } from "../../simple-tree/toFlexSchema.js";
+import { toStoredSchema } from "../../simple-tree/toStoredSchema.js";
 
 describe("toFlexSchema", () => {
 	it("minimal", () => {
